Filter and NativeQuery

+++ unit06-spring/04-spring-boot-backend-apirest/src/main/java/com/example/springboot/backend/apirest/models/services/UsuarioServiceImpl.java
@@ -45,6 +45,15 @@ public class UsuarioServiceImpl implements IUsuarioService {
 	@Override
 	@Transactional
 	public void delete(Long id) {
-		usuarioDao.deleteById(id);
+		Usuario u = usuarioDao.findById(id).orElse(null);
+		
+		if(u != null)
+		{
+			for(Mail m: u.getMails())
+			{
+				mailDao.delete(m);
+			}
+			usuarioDao.deleteById(id);
+		}
 	}
 }
